edit jokesTableSkeleton
