+++ components/jokes/edit-joke.tsx
@@ -13,7 +13,7 @@ import { UseCase } from "@/types/jokes-form"
 import { convertKeysToPascalCase } from "@/lib/utils"
 import { Button, buttonVariants } from "@/components/ui/button"
 import { Card, CardContent } from "@/components/ui/card"
-import { JokesFormSkeleton } from "../loading-skeletons/jokes-form"
+import { JokesFormSkeleton } from "../loading-skeletons/jokes"
 import {
   Dialog,
   DialogContent,
